Feature/enhace UI and dicovery (#18)
* test(e2e): validate admin registry-to-index flow

Add admin-focused E2E coverage for registry creation, trigger, and index verification with two query channels and query normalization in test calls. Add frontend admin route guard/login redirect support. Fix worker GitHub clone reliability by retrying public clone without token on auth errors, and harden code-search repo deserialization defaults.

* chore(dev): align refresh token schema and docker workflow

Align refresh token entity/auth types with migration, keep token insert error details in server logs, and add Docker workflow docs plus Makefile shortcuts. Update compose stack to use shared backend runtime image and expose Temporal port for local operations.

* Update skills

* feat(auth): support multiple frontend origins

* feat: add configurable discovery registry API URL

* ui update

+++ backend/common/src/entities/refresh_tokens.rs
@@ -5,11 +5,11 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(table_name = "refresh_tokens")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i64,
+    pub id: i32,
     pub user_id: Uuid,
     pub token_hash: String,
     #[sea_orm(nullable)]
-    pub rotated_from: Option<i64>,
+    pub rotated_from: Option<i32>,
     pub expires_at: DateTime,
     #[sea_orm(nullable)]
     pub revoked_at: Option<DateTime>,
